+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- check if simulation SLA contains `pickupPointId` from the selected pickup point
+
 ## [3.0.10] - 2019-09-12
 
 ### Fixed